llm_query_app_minimal_codespaces/requirements.txt
flask==2.3.3
google-generativeai==0.3.2
loguru==0.7.2
requests==2.31.0
python-dotenv==1.0.1
pyyaml==6.0.1
